+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+#### Changed
+- Use email as identifier instead of username
+
 
 ## [v1.0.0]
 - First release