+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+## [v1.1.0]
 #### Added
 - Added possibility to recover LS Hostel password
 #### Changed
@@ -12,4 +14,5 @@ All notable changes to this project will be documented in this file.
 - First release
  
 [Unreleased]: https://github.com/CESNET/lshostel-aai-proxy-idp-template/tree/master
+[v1.1.0]: https://github.com/CESNET/lshostel-aai-proxy-idp-template/tree/v1.1.0
 [v1.0.0]: https://github.com/CESNET/lshostel-aai-proxy-idp-template/tree/v1.0.0